LPS-76075 Move configurations that should not be exported to internal package leaving the same id as before so the previous configurations are kept
