+++ modules/apps/knowledge-base/knowledge-base-web/src/main/java/com/liferay/knowledge/base/web/internal/configuration/definition/KBDisplayPortletInstanceConfigurationBeanDeclaration.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.knowledge.base.web.configuration.definition;
+package com.liferay.knowledge.base.web.internal.configuration.definition;
 
-import com.liferay.knowledge.base.web.configuration.KBSectionPortletInstanceConfiguration;
+import com.liferay.knowledge.base.web.internal.configuration.KBDisplayPortletInstanceConfiguration;
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
 
 import org.osgi.service.component.annotations.Component;
@@ -23,12 +23,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Roberto Díaz
  */
 @Component
-public class KBSectionPortletInstanceConfigurationBeanDeclaration
+public class KBDisplayPortletInstanceConfigurationBeanDeclaration
 	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return KBSectionPortletInstanceConfiguration.class;
+		return KBDisplayPortletInstanceConfiguration.class;
 	}
 
 }
